Create factory files and db seeder file

+++ backend/database/migrations/2024_05_25_130446_create_registrations_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('registrations', function (Blueprint $table) {
             $table->id();
             
-            $table->unsignedTinyInteger("school_year");
+            $table->unsignedSmallInteger("school_year");
 
             $table->unsignedBigInteger("student_id");
             $table->unsignedBigInteger("group_id");
