+++ backend/database/migrations/2024_05_25_132532_create_absence_reports_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
 
             $table->string("reason", 100)->nullable();
-            $table->unsignedTinyInteger("school_year");
+            $table->unsignedSmallInteger("school_year");
             $table->enum("quarter", [1, 2, 3]);
 
             $table->unsignedBigInteger("student_id");
